LibWasm: Reduce StructureInstructionArgs' meta parameter size to 32b
Even this is overkill, but 64b fits in a register so no need to reduce
it further.

+++ Libraries/LibWasm/AbstractMachine/Validator.cpp
@@ -2108,8 +2108,8 @@ VALIDATE_INSTRUCTION(block)
         stack.append(parameter);
 
     args.meta = Instruction::StructuredInstructionArgs::Meta {
-        .arity = block_type.results().size(),
-        .parameter_count = parameters.size(),
+        .arity = static_cast<u32>(block_type.results().size()),
+        .parameter_count = static_cast<u32>(parameters.size()),
     };
 
     return {};
@@ -2130,8 +2130,8 @@ VALIDATE_INSTRUCTION(loop)
         stack.append(parameter);
 
     args.meta = Instruction::StructuredInstructionArgs::Meta {
-        .arity = block_type.results().size(),
-        .parameter_count = parameters.size(),
+        .arity = static_cast<u32>(block_type.results().size()),
+        .parameter_count = static_cast<u32>(parameters.size()),
     };
 
     return {};
@@ -2156,8 +2156,8 @@ VALIDATE_INSTRUCTION(if_)
         stack.append(parameter);
 
     args.meta = Instruction::StructuredInstructionArgs::Meta {
-        .arity = block_type.results().size(),
-        .parameter_count = parameters.size(),
+        .arity = static_cast<u32>(block_type.results().size()),
+        .parameter_count = static_cast<u32>(parameters.size()),
     };
 
     return {};
@@ -2207,8 +2207,8 @@ VALIDATE_INSTRUCTION(try_table)
         TRY(stack.take(parameters[parameters.size() - i]));
 
     args.try_.meta = Instruction::StructuredInstructionArgs::Meta {
-        .arity = block_type.results().size(),
-        .parameter_count = parameters.size(),
+        .arity = static_cast<u32>(block_type.results().size()),
+        .parameter_count = static_cast<u32>(parameters.size()),
     };
 
     m_frames.empend(block_type, FrameKind::TryTable, stack.size());
